refactor(backend): Change the directory structure
Move cocktail API to the `router` directory.
