+++ backend/src/api/router/cocktail.ts
@@ -1,0 +1,40 @@
+import { type Context, Hono } from "@hono/hono";
+export type { Hono };
+import { STATUS_CODE } from "@std/http/status";
+
+import { Cocktail } from "../utils/types.ts";
+export { Cocktail };
+import { cocktails } from "../utils/data.ts";
+
+/**
+ * The cocktail API
+ *
+ * @example Return the requested cocktail detail
+ * ```ts
+ * const res: Response = await api.request("/?name=アイリッシュコーヒー");
+ * ```
+ * @example Return an array of all cocktails
+ * ```ts
+ * const res: Response = await api.request("/all");
+ * ```
+ */
+export const app: Hono = new Hono();
+app
+  .get("/", (ctx: Context) => {
+    const name: string | undefined = ctx.req.query("name");
+    if (!name) {
+      return ctx.json(
+        { success: false, message: 'The "name" query is required' },
+        STATUS_CODE.BadRequest,
+      );
+    }
+
+    const cocktail: Cocktail | undefined = cocktails.find((c: Cocktail) =>
+      c.name === name
+    );
+    return cocktail ? ctx.json({ success: true, data: cocktail }) : ctx.json(
+      { success: false, message: `The cocktail, "${name}" not found` },
+      STATUS_CODE.NotFound,
+    );
+  })
+  .get("/all", (ctx: Context) => ctx.json(cocktails));
