+++ backend/src/api/router/mod.ts
@@ -1,0 +1,1 @@
+export { app as cocktailApi } from "./cocktail.ts";
