+++ backend/src/server.test.ts
@@ -7,7 +7,7 @@ Deno.test("Serve", async (t: Deno.TestContext) => {
   await t.step("GET /", async () => {
     const res: Response = await app.request("/");
 
-    assertEquals(await res.text(), "Hello, World!");
+    assertEquals(await res.text(), "Cocktail Message App");
     assertEquals(res.status, STATUS_CODE.OK);
   });
 });
